+ setup-ubuntu.sh

+++ setup-ubuntu.sh
@@ -10,7 +10,9 @@ HDIR=/mnt/data/song10
 [ -d $HDIR ] || exit 1
 rm -f examples.desktop
 for x in Desktop Documents Downloads Music Pictures Public Templates Videos; do
-	rmdir $x
 	[ -d $HDIR/$x ] || mkdir -p $HDIR/$x
-	ln -s $HDIR/$x
+	if [ ! -s $x ]; then
+		rmdir $x # --ignore-fail-on-non-empty
+		ln -s $HDIR/$x
+	fi
 done
